requirements.txt
Flask==2.3.3
transformers==4.40.1
torch==2.3.0
sentencepiece==0.1.99
accelerate==0.30.1
pymongo==4.6.3
dnspython==2.4.2
python-dotenv 